Add files via upload
This study aims to devised another vaccination method. A web-based vaccination scheduling system to make it easier for people to get vaccinated, as we are avoiding crowded areas now that there is a pandemic. Also, to provide easier access for people who can't reach out directly to barangay bookings. In addition, with the vaccination schedule, the system will be distinctive to others by generating a unique QR code for each registrant that will serve as a copy of their vaccination card and information.